hotfix: complete cast of alimage into ros::time

+++ naoqi_sensors/src/naoqi_camera.cpp
@@ -373,7 +373,7 @@ namespace naoqicamera_driver
             image->header.stamp = ros::Time::now();
         else { 
             // use NAOqi timestamp
-            image->header.stamp = ros::Time(al_image[4], al_image[5] * 1000);
+            image->header.stamp = ros::Time(static_cast<int>(al_image[4]), static_cast<int>(al_image[5]) * 1000);
         }
 
         image->width = (int) al_image[0];
